refa: Dao 역할 분리
