+++ src/main/java/repository/game/JdbcGameDao.java
@@ -1,55 +1,21 @@
-package repository;
+package repository.game;
 
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.sql.Statement;
 import java.util.ArrayList;
 import java.util.List;
 
 import dto.BoardDto;
 import dto.MoveHistoryDto;
+import repository.connector.JdbcConnector;
 
-public class JdbcChessDao implements ChessDao {
-
+public class JdbcGameDao implements GameDao{
     private final JdbcConnector connector;
 
-    public JdbcChessDao(JdbcConnector connector) {
+    public JdbcGameDao(JdbcConnector connector) {
         this.connector = connector;
-    }
-
-    @Override
-    public long addGame(String gameName) {
-        final String query = "INSERT INTO game (gameName) VALUES (?)";
-        try (Connection connection = connector.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(query,
-                     Statement.RETURN_GENERATED_KEYS)) {
-            preparedStatement.setString(1, gameName);
-            preparedStatement.executeUpdate();
-            ResultSet generatedKeys = preparedStatement.getGeneratedKeys();
-            generatedKeys.next();
-            return generatedKeys.getLong(1);
-        } catch (final SQLException exception) {
-            throw new RuntimeException(exception);
-        }
-    }
-
-    @Override
-    public List<String> findAllGame() {
-        final String query = "SELECT gameName FROM game";
-        try (Connection connection = connector.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
-            ResultSet resultSet = preparedStatement.executeQuery();
-            ArrayList<String> gameNames = new ArrayList<>();
-            while (resultSet.next()) {
-                String gameName = resultSet.getString("gameName");
-                gameNames.add(gameName);
-            }
-            return gameNames;
-        } catch (SQLException exception) {
-            throw new RuntimeException(exception);
-        }
     }
 
     @Override
@@ -69,17 +35,6 @@ public class JdbcChessDao implements ChessDao {
                 boardDtos.add(new BoardDto(square, piece, camp));
             }
             return boardDtos;
-        } catch (SQLException exception) {
-            throw new RuntimeException(exception);
-        }
-    }
-
-    @Override
-    public void deleteAllGame() {
-        final String query = "DELETE from game";
-        try (Connection connection = connector.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
-            preparedStatement.executeUpdate();
         } catch (SQLException exception) {
             throw new RuntimeException(exception);
         }
@@ -119,16 +74,13 @@ public class JdbcChessDao implements ChessDao {
     }
 
     @Override
-    public long findGameIdByGameName(String gameName) {
-        final String query = "SELECT _id FROM game WHERE gameName = ?";
+    public void updateCurrentTurn(long gameId, String currentTurn) {
+        final String query = "UPDATE game SET currentTurn = ? WHERE _id = ?";
         try (Connection connection = connector.getConnection();
              PreparedStatement preparedStatement = connection.prepareStatement(query)) {
-            preparedStatement.setString(1, gameName);
-            ResultSet resultSet = preparedStatement.executeQuery();
-            if (resultSet.next()) {
-                return resultSet.getInt("_id");
-            }
-            throw new IllegalArgumentException("존재하지 않는 게임입니다.");
+            preparedStatement.setString(1, currentTurn);
+            preparedStatement.setLong(2, gameId);
+            preparedStatement.executeUpdate();
         } catch (SQLException exception) {
             throw new RuntimeException(exception);
         }
@@ -149,19 +101,6 @@ public class JdbcChessDao implements ChessDao {
     }
 
     @Override
-    public void updateCurrentTurn(long gameId, String currentTurn) {
-        final String query = "UPDATE game SET currentTurn = ? WHERE _id = ?";
-        try (Connection connection = connector.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
-            preparedStatement.setString(1, currentTurn);
-            preparedStatement.setLong(2, gameId);
-            preparedStatement.executeUpdate();
-        } catch (SQLException exception) {
-            throw new RuntimeException(exception);
-        }
-    }
-
-    @Override
     public void deleteBoardById(long gameId) {
         final String query = "DELETE from board`` WHERE g_id = ?";
         try (Connection connection = connector.getConnection();
@@ -173,6 +112,7 @@ public class JdbcChessDao implements ChessDao {
         }
     }
 
+    @Override
     public List<MoveHistoryDto> findMoveHistoryByGameId(long game_id) {
         final String query = "SELECT source, target, pieceOnTarget FROM moveHistory WHERE g_id = ?";
         try (Connection connection = connector.getConnection();
@@ -187,27 +127,6 @@ public class JdbcChessDao implements ChessDao {
                 moveHistoryDtos.add(new MoveHistoryDto(source, target, piece));
             }
             return moveHistoryDtos;
-        } catch (SQLException exception) {
-            throw new RuntimeException(exception);
-        }
-    }
-
-
-    public void deleteAllBoard() {
-        final String query = "DELETE from board";
-        try (Connection connection = connector.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
-            preparedStatement.executeUpdate();
-        } catch (SQLException exception) {
-            throw new RuntimeException(exception);
-        }
-    }
-
-    public void deleteAllMoveHistory() {
-        final String query = "DELETE from moveHistory";
-        try (Connection connection = connector.getConnection();
-             PreparedStatement preparedStatement = connection.prepareStatement(query)) {
-            preparedStatement.executeUpdate();
         } catch (SQLException exception) {
             throw new RuntimeException(exception);
         }
