+++ src/test/java/repository/game/JdbcGameDaoTest.java
@@ -1,55 +1,32 @@
-package repository;
+package repository.game;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import java.sql.Connection;
-import java.sql.SQLException;
 import java.util.List;
 
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
 import dto.BoardDto;
 import dto.MoveHistoryDto;
+import repository.TestConnector;
+import repository.connector.JdbcConnector;
+import repository.room.JdbcRoomDao;
 
-class JdbcChessDaoTest {
+class JdbcGameDaoTest {
+
     JdbcConnector connector = new TestConnector();
-    JdbcChessDao jdbcChessDao = new JdbcChessDao(connector);
-
-    @BeforeEach
-    void setUp() {
-        jdbcChessDao.deleteAllMoveHistory();
-        jdbcChessDao.deleteAllBoard();
-        jdbcChessDao.deleteAllGame();
-    }
-
-    @Test
-    @DisplayName("DB 커넥션 테스트")
-    void connection() throws SQLException {
-        try (Connection connection = connector.getConnection()) {
-            assertThat(connection).isNotNull();
-        }
-    }
-
-    @Test
-    @DisplayName("addGame을 통해 DB에 게임을 저장한다.")
-    void addGame() {
-        jdbcChessDao.addGame("테스트");
-
-        List<String> allGame = jdbcChessDao.findAllGame();
-
-        assertThat(allGame).contains("테스트");
-    }
+    JdbcRoomDao jdbcChessDao = new JdbcRoomDao(connector);
+    JdbcGameDao jdbcGameDao = new JdbcGameDao(connector);
 
     @Test
     @DisplayName("saveMoveHistory를 통해 move를 저장한다.")
     void saveMoveHistory() {
-        long gameId = jdbcChessDao.addGame("테스트");
+        long gameId = jdbcChessDao.createRoom("테스트");
         MoveHistoryDto moveHistoryDto = new MoveHistoryDto("a2", "a3", "EMPTY");
-        jdbcChessDao.saveMoveHistory(gameId, moveHistoryDto);
+        jdbcGameDao.saveMoveHistory(gameId, moveHistoryDto);
 
-        List<MoveHistoryDto> moveHistoryByGameId = jdbcChessDao.findMoveHistoryByGameId(gameId);
+        List<MoveHistoryDto> moveHistoryByGameId = jdbcGameDao.findMoveHistoryByGameId(gameId);
 
         assertThat(moveHistoryByGameId).contains(moveHistoryDto);
     }
@@ -57,7 +34,7 @@ class JdbcChessDaoTest {
     @Test
     @DisplayName("saveBoard를 통해 현재보드를 저장한다.")
     void saveBoard() {
-        long gameId = jdbcChessDao.addGame("테스트");
+        long gameId = jdbcChessDao.createRoom("테스트");
         List<BoardDto> boardDtos = List.of(
                 new BoardDto("a2", "PAWN", "WHITE"),
                 new BoardDto("a3", "QUEEN", "BLACK"),
@@ -67,8 +44,8 @@ class JdbcChessDaoTest {
                 new BoardDto("a7", "KING", "BLACK")
         );
 
-        jdbcChessDao.saveBoard(gameId, boardDtos);
-        List<BoardDto> boardDto = jdbcChessDao.findBoardByGameName("테스트");
+        jdbcGameDao.saveBoard(gameId, boardDtos);
+        List<BoardDto> boardDto = jdbcGameDao.findBoardByGameName("테스트");
 
         assertThat(boardDto).containsExactlyInAnyOrderElementsOf(boardDto);
     }
@@ -76,7 +53,7 @@ class JdbcChessDaoTest {
     @Test
     @DisplayName("gameId를 전달 받아 해당하는 board를 삭제한다.")
     void deleteBoardById() {
-        long gameId = jdbcChessDao.addGame("테스트");
+        long gameId = jdbcChessDao.createRoom("테스트");
         List<BoardDto> boardDtos = List.of(
                 new BoardDto("a2", "PAWN", "WHITE"),
                 new BoardDto("a3", "QUEEN", "BLACK"),
@@ -85,11 +62,11 @@ class JdbcChessDaoTest {
                 new BoardDto("a6", "KING", "WHITE"),
                 new BoardDto("a7", "KING", "BLACK")
         );
-        jdbcChessDao.saveBoard(gameId, boardDtos);
+        jdbcGameDao.saveBoard(gameId, boardDtos);
 
-        jdbcChessDao.deleteBoardById(gameId);
+        jdbcGameDao.deleteBoardById(gameId);
 
-        List<BoardDto> boardDto = jdbcChessDao.findBoardByGameName("테스트");
+        List<BoardDto> boardDto = jdbcGameDao.findBoardByGameName("테스트");
         assertThat(boardDto).isEmpty();
     }
 }
